--- src/app/mypage/page.tsx
@@ -1,13 +1,0 @@
-import MyPageProfile from "@/components/mypage/MyPageProfile";
-import React from "react";
-
-const Mypage = () => {
-  return (
-    <div>
-      <h1>마이페이지</h1>
-      <MyPageProfile />
-    </div>
-  );
-};
-
-export default Mypage;
